+++ src/main/java/com/project2/wanderfun/application/service/PlaceService.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface PlaceService extends BaseService<Place>{
     public List<Place> findAllByNameContaining(String name);
     public Place findByName(String name);
-    public Place findByLongitudeAndLatitude(BigDecimal longitude, BigDecimal latitude);
+    public Place findByLongitudeAndLatitude(double longitude, double latitude);
 }
